# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Shradha Oil Center ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all accounts
      responses:
        "200":
          description: Accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    put:
      operationId: updateAccount
      tags: [accounts]
      summary: Update an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete (deactivate) an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # CASH BOOK — salesman cash collection
...
  /account-transactions:
    get:
      operationId: listAccountTransactions
      tags: [accounts]
      summary: List account transactions (cash in/out)
      parameters:
        - name: accountId
          in: query
          schema:
            type: integer
        - name: direction
          in: query
          schema:
            type: string
            enum: [in, out]
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountTransaction"
    post:
      operationId: createAccountTransaction
      tags: [accounts]
      summary: Record a Payment In or Payment Out against an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountTransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountTransaction"
  /cashbook/collect:
    post:
      operationId: collectCashFromSalesman
      tags: [accounts]
      summary: Record cash collected from a salesman and deposit into an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectCashInput"
      responses:
        "200":
          description: Collected
          content:
            application/json:
              schema:
                type: object
                required: [collectedCount, totalAmount, account]
                properties:
                  collectedCount:
                    type: integer
                  totalAmount:
                    type: number
                  account:
                    $ref: "#/components/schemas/Account"

  # REWARD SCHEMES
...
components:
  schemas:
...
    Account:
      type: object
      required: [id, name, type, openingBalance, currentBalance, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [cash, upi, bank]
        identifier:
          type: ["string", "null"]
        openingBalance:
          type: number
        currentBalance:
          type: number
        isActive:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    AccountInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [cash, upi, bank]
        identifier:
          type: string
        openingBalance:
          type: number
        notes:
          type: string
        isActive:
          type: boolean
...
    CollectCashInput:
      type: object
      required: [salesmanId, accountId, amount]
      properties:
        salesmanId:
          type: integer
        accountId:
          type: integer
        amount:
          type: number
        notes:
          type: string
    AccountTransaction:
      type: object
      required: [id, accountId, direction, amount, mode, createdAt]
      properties:
        id:
          type: integer
        receiptNo:
          type: ["string", "null"]
        accountId:
          type: integer
        accountName:
          type: ["string", "null"]
        direction:
          type: string
          enum: [in, out]
        amount:
          type: number
        mode:
          type: string
          enum: [cash, upi, bank_transfer, cheque, other]
        partyName:
          type: ["string", "null"]
        partyMobile:
          type: ["string", "null"]
        partyEntityId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdById:
          type: ["integer", "null"]
        createdByName:
          type: ["string", "null"]
        createdByRole:
          type: ["string", "null"]
        balanceAfter:
          type: ["number", "null"]
        createdAt:
          type: string
    AccountTransactionInput:
      type: object
      required: [accountId, direction, amount, mode]
      properties:
        accountId:
          type: integer
        direction:
          type: string
          enum: [in, out]
        amount:
          type: number
          minimum: 0.01
        mode:
          type: string
          enum: [cash, upi, bank_transfer, cheque, other]
        partyName:
          type: string
        partyMobile:
          type: string
        partyEntityId:
          type: integer
        notes:
          type: string
...